backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pymongo==4.8.0
requests==2.32.3
pypdf==5.0.1
gunicorn==23.0.0
psycopg[binary]==3.2.1
psycopg-pool==3.2.3
redis==5.2.0
rq==1.16.2
stripe==10.12.0
